fix: now duiring on showPersonFragment person doesnt deleting now

+++ app/src/main/java/com/example/atosrm/presentation/fr/show_full_info_person/ShowPersonSRMFragment.kt
@@ -174,8 +174,10 @@ import okhttp3.internal.isSensitiveHeader
     LaunchedEffect(
         key1 = deletePersonTrigger,
         block = {
-            dao.deletePerson(person = person)
-            navController.popBackStack()
+            if(deletePersonTrigger > 0 ) {
+                dao.deletePerson(person = person)
+                navController.popBackStack()
+            }
         }
     )
